Restyle README
Completely redesign README.

Method 2 for manual installation needs work for Firefox. Installing with the standard method results in an error, as I would imagine the build info is missing from the repo, or there is a method I am missing.

Open to discussion on this, and a resubmit if necessary.